remove usize and add indicator foundations

+++ ff_standard_lib/src/consolidators/consolidator_enum.rs
@@ -124,7 +124,7 @@ impl ConsolidatorEnum {
         }
     }
 
-    pub fn index(&self, index: usize) -> Option<BaseDataEnum> {
+    pub fn index(&self, index: u64) -> Option<BaseDataEnum> {
         match self {
             ConsolidatorEnum::Count(count_consolidator) => count_consolidator.index(index),
             ConsolidatorEnum::CandleStickConsolidator(time_consolidator) => time_consolidator.index(index),
